Fix ABI after moving types from shared crate
