+++ Cargo.toml
@@ -12,7 +12,7 @@ repository = "https://github.com/slowli/tardigrade"
 
 [dependencies]
 async-trait = "0.1.57"
-chrono = { version = "0.4.22", default-features = false }
+chrono = { version = "0.4.22", default-features = false, features = ["serde"] }
 futures = "0.3.21"
 once_cell = "1.12.0"
 pin-project-lite = "0.2.9"
