--- src/types.rs
@@ -1,24 +1,0 @@
-//! Types shared between host and client envs.
-
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-
-/// ID of a [`Waker`](std::task::Waker) defined by a workflow.
-pub type WakerId = u64;
-/// ID of a workflow task.
-pub type TaskId = u64;
-/// ID of a workflow timer.
-pub type TimerId = u64;
-/// ID of a (traced) future defined by a workflow.
-pub type FutureId = u64;
-/// ID of a workflow.
-pub type WorkflowId = u64;
-/// ID of a channel.
-pub type ChannelId = u128;
-
-/// Definition of a timer used by a workflow.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct TimerDefinition {
-    /// Expiration timestamp of the timer.
-    pub expires_at: DateTime<Utc>,
-}
